requirements.txt
apache-airflow==2.10.0
pyspark==3.3.2
pymongo==4.3.3
minio==7.1.14
apache-airflow-providers-apache-spark==5.0.0